DW-3420: Fix Concourse-Cognito integration
When trying to login to Concourse via the Cognito integration, I was seeing a
`invalid_scope` error. It turns out that the `profile` scope is required and
the `aws.cognito.signin.user.admin` isn't; at least for the basic login flow
to work.

Also, remove the `explicit_auth_flows` as they're classed as optional and
omitting them doesn't seem to break anything...yet.